권한이 있는 유저만 유저 목표에 접근 가능하도록 수정

+++ src/main/java/com/modo/modoapiserver/controller/usergoal/UserGoalController.java
@@ -87,9 +87,14 @@ public class UserGoalController {
 
     @Operation(summary = "목표 상세 가져오기", description = "목표 상세를 가져옵니다")
     @GetMapping("/goal/{id}")
-    public ResponseEntity<UserGoalDto> getUserGoal(@PathVariable("id") Long id){
-        // TODO: jwt 토큰에서 사용자 정보를 가져와서 사용자의 목표인지 확인하는 로직 추가
+    public ResponseEntity<UserGoalDto> getUserGoal(@PathVariable("id") Long id, @AuthenticationPrincipal CustomUserDetails userDetails){
         UserGoal userGoal = userGoalService.getUserGoal(id);
+
+        boolean isUserGoalAccessible = userGoalService.isUserGoalAccessible(id, userDetails);
+
+        if(!isUserGoalAccessible) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "해당 목표에 접근할 권한이 없습니다.");
+        }
 
         Date expiration = new Date(System.currentTimeMillis() + 1000 * 60 * 60);
 
@@ -98,7 +103,6 @@ public class UserGoalController {
                 userGoal.getCompleteVerificationPictureUrl(),
                 expiration
         );
-
 
         UserGoalDto userGoalDto = UserGoalDto.builder()
                 .id(userGoal.getId())
@@ -119,17 +123,30 @@ public class UserGoalController {
 
     @Operation(summary = "목표 지우기", description = "목표를 삭제합니다.")
     @DeleteMapping("/goal/{id}")
-    public ResponseEntity<?> deleteUserGoal(@PathVariable("id") Long id){
-        // TODO: jwt 토큰에서 사용자 정보를 가져와서 사용자의 목표인지 확인하는 로직 추가
+    public ResponseEntity<?> deleteUserGoal(@PathVariable("id") Long id, @AuthenticationPrincipal CustomUserDetails userDetails){
+        boolean isUserGoalAccessible = userGoalService.isUserGoalAccessible(id, userDetails);
+
+        if(!isUserGoalAccessible) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "해당 목표에 접근할 권한이 없습니다.");
+        }
+
         userGoalService.deleteUserGoal(id);
+
         return ResponseEntity.ok().build();
     }
 
     @Operation(summary = "목표 업데이트", description = "목표를 업데이트합니다.")
     @PatchMapping("/goal/{id}")
-    public ResponseEntity<?> patchUserGoal(@PathVariable("id") Long id, @RequestBody RequestUserGoalDto requestUserGoalDto){
+    public ResponseEntity<?> patchUserGoal(@PathVariable("id") Long id, @RequestBody RequestUserGoalDto requestUserGoalDto, @AuthenticationPrincipal CustomUserDetails userDetails){
         // TODO: jwt 토큰에서 사용자 정보를 가져와서 사용자의 목표인지 확인하는 로직 추가
         UserGoal userGoal = userGoalService.getUserGoal(id);
+
+        boolean isUserGoalAccessible = userGoalService.isUserGoalAccessible(id, userDetails);
+
+        if(!isUserGoalAccessible) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "해당 목표에 접근할 권한이 없습니다.");
+        }
+
         if( requestUserGoalDto.getTitle() != null) {
             userGoal.setTitle(requestUserGoalDto.getTitle());
         }
@@ -217,6 +234,12 @@ public class UserGoalController {
             @RequestPart("file") MultipartFile file
     ) throws IOException {
         UserGoal userGoal = userGoalService.getUserGoal(id);
+
+        boolean isUserGoalAccessible = userGoalService.isUserGoalAccessible(id, userDetails);
+
+        if(!isUserGoalAccessible) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "해당 목표에 접근할 권한이 없습니다.");
+        }
 
         String bucketName = "modo-static-bucket";
         String fileExtension = file.getOriginalFilename().substring(file.getOriginalFilename().lastIndexOf("."));
